requirements.txt
biopython
keras
numpy
tensorflow
scikit-learn
tqdm
pyyaml
